+++ H264Decoder/Data/NALU/DPSRbsp.h
@@ -8,22 +8,22 @@ struct DecodingContext;
 
 struct DPSRbsp
 {
-    std::uint8_t dpsId = 0;
-    std::uint8_t predDirection = 0;
+    std::uint16_t dpsId = 0;
+    std::uint16_t predDirection = 0;
 
-    std::uint8_t refDpsId0 = 0;
-    std::uint8_t refDpsId1 = 0;
+    std::uint16_t refDpsId0 = 0;
+    std::uint16_t refDpsId1 = 0;
     std::uint8_t predWeight0 = 0;
 
-    std::uint8_t numDepthViewsMinus1 = 0;
+    std::uint16_t numDepthViewsMinus1 = 0;
     DepthRange depthRanges;
 
     bool vspParamFlag = false;
     VspParam vspParam;
 
     bool depthParamAdditionalExtensionFlag = false;
-    std::uint8_t nonlinearDepthRepresentationNum = 0;
-    std::vector<std::uint8_t> nonlinearDepthRepresentationModel;
+    std::uint16_t nonlinearDepthRepresentationNum = 0;
+    std::vector<std::uint16_t> nonlinearDepthRepresentationModel;
 
     std::vector<BoolType> depthParamAdditionalExtensionDataFlags;
